Fix import errors from file rename.
